functions/requirements.txt
firebase_functions~=0.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
